requirements.txt
antlr4-python3-runtime==4.9.3
certifi==2022.12.7
charset-normalizer==3.0.1
hydra-core==1.3.1
idna==3.4
imageio==2.25.1
importlib-resources==5.12.0
numpy==1.24.2
omegaconf==2.3.0
packaging==23.0
Pillow==9.4.0
PyYAML==6.0
requests==2.28.2
torch==1.11.0+cu113
torchaudio==0.11.0+cu113
torchvision==0.12.0+cu113
tqdm==4.64.1
typing-extensions==4.5.0
urllib3==1.26.14
zipp==3.14.0
